+++ index.js
@@ -1,9 +1,9 @@
-const { version } = require('discord.js');
-let ModalSubmitInteraction = require('./src/structures/ModalSubmitInteraction');
-const { Error } = require('./src/structures/errors');
-const { InteractionTypes } = require('./src/util/Constants');
+const { version } = require("discord.js");
+let ModalSubmitInteraction = require("./src/structures/ModalSubmitInteraction");
+const { Error } = require("./src/structures/errors");
+const { InteractionTypes } = require("./src/util/Constants");
 const discordjsVersion = new String('v' + version); // Expected: v13 or v14...
-if(discordjsVersion.includes('v14')) ModalSubmitInteraction = require('./src/v14/ModalSubmitInteraction');
+if(discordjsVersion.includes('v14')) ModalSubmitInteraction = require("./src/v14/ModalSubmitInteraction");
 
 module.exports = (client) => {
 
@@ -35,28 +35,28 @@ module.exports = (client) => {
 
 // Exports the classes according to the discord.js version.
 
-if (discordjsVersion.includes('v13')) {
-  module.exports.Modal = require('./src/structures/Modal');
-  module.exports.TextInputComponent = require('./src/structures/TextInputComponent');
-  module.exports.ModalSubmitInteraction = require('./src/structures/ModalSubmitInteraction');
-  module.exports.ModalSubmitField = require('./src/structures/ModalSubmitField');
-  module.exports.ModalActionRow = require('./src/structures/ModalActionRow');
-  module.exports.showModal = require('./src/structures/ShowModal');
-  module.exports.Interaction = require('./src/structures/Interaction');
-  module.exports.InteractionResponses = require('./src/structures/interfaces/InteractionResponses');
-  module.exports.Constants = require('./src/util/Constants');
-  module.exports.SnowflakeUtil = require('./src/util/SnowflakeUtil');
-} else if (discordjsVersion.includes('v14')) {
-  module.exports.Modal = require('./src/structures/Modal');
-  module.exports.TextInputComponent = require('./src/structures/TextInputComponent');
-  module.exports.ModalSubmitInteraction = require('./src/v14/ModalSubmitInteraction');
-  module.exports.ModalSubmitField = require('./src/structures/ModalSubmitField');
-  module.exports.ModalActionRow = require('./src/structures/ModalActionRow');
-  module.exports.showModal = require('./src/v14/ShowModal');
-  module.exports.Interaction = require('./src/v14/Interaction');
-  module.exports.InteractionResponses = require('./src/v14/interfaces/InteractionResponses');
-  module.exports.Constants = require('./src/util/Constants');
-  module.exports.SnowflakeUtil = require('./src/util/SnowflakeUtil');
+if (discordjsVersion.includes("v13")) {
+  module.exports.Modal = require("./src/structures/Modal");
+  module.exports.TextInputComponent = require("./src/structures/TextInputComponent");
+  module.exports.ModalSubmitInteraction = require("./src/structures/ModalSubmitInteraction");
+  module.exports.ModalSubmitField = require("./src/structures/ModalSubmitField");
+  module.exports.ModalActionRow = require("./src/structures/ModalActionRow");
+  module.exports.showModal = require("./src/structures/ShowModal");
+  module.exports.Interaction = require("./src/structures/Interaction");
+  module.exports.InteractionResponses = require("./src/structures/interfaces/InteractionResponses");
+  module.exports.Constants = require("./src/util/Constants");
+  module.exports.SnowflakeUtil = require("./src/util/SnowflakeUtil");
+} else if (discordjsVersion.includes("v14")) {
+  module.exports.Modal = require("./src/structures/Modal");
+  module.exports.TextInputComponent = require("./src/structures/TextInputComponent");
+  module.exports.ModalSubmitInteraction = require("./src/v14/ModalSubmitInteraction");
+  module.exports.ModalSubmitField = require("./src/structures/ModalSubmitField");
+  module.exports.ModalActionRow = require("./src/structures/ModalActionRow");
+  module.exports.showModal = require("./src/v14/ShowModal");
+  module.exports.Interaction = require("./src/v14/Interaction");
+  module.exports.InteractionResponses = require("./src/v14/interfaces/InteractionResponses");
+  module.exports.Constants = require("./src/util/Constants");
+  module.exports.SnowflakeUtil = require("./src/util/SnowflakeUtil");
 }
 
 /* Powered by:
@@ -67,5 +67,4 @@ if (discordjsVersion.includes('v13')) {
 ------ Developed by 『𝑴𝒂𝒕𝒆𝒐ᵗᵉᵐ』#9999 ------
 
 https://www.npmjs.com/package/discord-modals
-
 */