+++ src/structures/ModalSubmitField.js
@@ -1,15 +1,19 @@
-const BaseMessageComponent = require('./BaseMessageComponent');
-
 /**
  * Represents a Field of a Modal Submit Interaction.
- * @extends BaseMessageComponent
-*/
+ */
 
-class ModalSubmitField extends BaseMessageComponent {
+class ModalSubmitField {
   constructor(data = {}) {
-    super({ type: 'TEXT_INPUT' });
+
+    /**
+     * The type of the Modal Submit Field.
+     * @type {String}
+     */
+
+    this.type = 'TEXT_INPUT';
 
     this.setup(data);
+
   }
 
   setup(data) {
@@ -17,14 +21,14 @@ class ModalSubmitField extends BaseMessageComponent {
     /**
      * The Custom Id of the Modal Submit Field.
      * @type {String}
-    */
+     */
 
     this.customId = data.custom_id ?? null;
 
     /**
      * The Value of the Modal Submit Field.
      * @type {String}
-    */
+     */
 
     switch(data.value.trim().length) {
       case 0: