+++ src/v14/Interaction.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const { Base, PermissionsBitField } = require('discord.js');
-const SnowflakeUtil = require('../util/SnowflakeUtil');
-const { InteractionType } = require('discord-api-types/v10');
+const { Base, PermissionsBitField } = require("discord.js");
+const SnowflakeUtil = require("../util/SnowflakeUtil");
+const { InteractionType } = require("discord-api-types/v9");
 
 /**
  * Represents an Interaction.