+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 Hello! This is the Change log section of Discord-Modals. Here you can see what things have been changed in the different versions, to be informed. This will be updated every time there are new versions. I hope you enjoy this package :)
 
-# v1.3.1 (Stable)
+# v1.3.2 (Stable)
+
+- Compatibility: change `discord-api-types` enums from `v10` to `v9`.
+- `Modal` and `ModalSubmitInteraction` class: Now the `.components` property returns `ModalActionRow`s.
+- Code cleanup.
+
+# v1.3.1
 - Fix: change `discord-api-types` enums from `v9` to `v10`.
 
 # v1.3.0
 
 - Fix bug: 'INVALID_CLIENT' error when the Client is valid on `.showModal()` method.
-- **Featured:** Add `ModalActionRow` class.
 - Now the `TextInputComponent` returns an **Action Row Component**. Add `ModalActionRow` class and types.
 - **Featured:** Now the `.showModal()` method supports **JSON Modals**.
 - Fix error typings `ModalSubmitField`.